Camel case naming convention for variables ($tokenId)
    In the route you have not mentioned the appointmentCategory. I think it should be passed as query string. Is the appointmentCategory mandatory? If not, you need to check if the query string has appointmentCategory in the controller

    You should not pass the entire request to the service layer. Extract the appointmentCategory as variable from the request in the controller and then pass the variable to the service layer

    The exception error code (PATIENT_REPORTS_LIST_ERROR) that you are throwing in consistent and not relevant. I believe you have just copy and pasted the code. When you do copy & paste, please make sure you defined a proper error code

    Also, the error code that you catch in the service and the controller should be the same.

    Also, I do not find a relevant message defined for the error code in messages.php file. Please define proper and appropriate message

Regarding mandatory field validation for patient register, Sham has sent a mail with the list of mandatory fields. Please do not make it as specific and not like "remove as many mandatory fields as possible". You need to handle the validation both at the client side as well as server side. In the server side, you need to remove the validations from the Request also.

all above points was done